# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentação da API Home Server Hub",
        "title": "Home Server Hub API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
        "/applications": {
            "get": {
                "description": "Retorna todas as aplicações já criadas e armazenadas no sistema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Lista aplicações",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ListApplicationsResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/applications/": {
            "post": {
                "description": "Cria e armazena uma aplicação com base no ID de um container Docker e dados opcionais enviados via multipart/form-data",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Cria uma nova aplicação",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do container Docker",
                        "name": "container_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nome personalizado da aplicação",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "format": "int32",
                        "description": "Porta a ser exposta",
                        "name": "port",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "URL pública da aplicação",
                        "name": "url",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Imagem opcional para representar a aplicação",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Application"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/applications/discover": {
            "get": {
                "description": "Busca por containers Docker em execução e retorna como aplicações potenciais",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Descobre aplicações Docker rodando no servidor",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DiscoverResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/applications/{id}/image": {
            "get": {
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "applications"
                ],
                "summary": "Retorna a imagem de uma aplicação",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID da aplicação",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Application": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "image": {
                    "$ref": "#/definitions/models.Image"
                },
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "status": {
                    "description": "calculado em tempo real, não persistido",
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.DiscoverResult": {
            "type": "object",
            "properties": {
                "discovered": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DiscoveredApplication"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.DiscoveredApplication": {
            "type": "object",
            "properties": {
                "container": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.Image": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "models.ListApplicationsResult": {
            "type": "object",
            "properties": {
                "applications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Application"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}